No error if m-buffer is not present.
m-buffer is required with dir locals so that the macros indent properly.
Otherwise, I get lots of false diffs as the macro indentation changes
depending on whether m-buffer has been required. As m-buffer is required
by lentic, normally this will work, unless a random developer has
cloned lentic, but not actually installed it. In that case, there is no
reason to crash when m-buffer is not present.