lentic-mode now works, and has a lighter.
Lentic-mode actually did nothing and was never switched on. It has been
combined with lentic-start-mode which did not really have a purpose. It
now also has a lighter which works and which can be configured on using
the lentic-configuration object. By default, it just shows whether a
lentic buffer exists. lentic-block-configuration also uses this to
display whether the current-buffer is invalid wrt to its markers.