Force install on every test.
This is a bit of a pain as it slows down tests, requiring two emacs
launches. The advantage is that the test never fails with dependency
errors. This was happening a lot for me as I have several versions of
Emacs installed -- git appears to be running a different one on
pre-commit from the one I get at the shell.