lentic-doc now generic for any package.
Previously, large parts of this were hard-coded to lentic. Now
we only have two functions (which are used in lentic-mode) specific to
lentic itself. The rest should work for any package which has lenticular
documentation.

This required a way of specifying the main source file -- with lentic,
for example, this is lenticular.org. We achieve this by looking for a
specific variable, a standard name or just use the main package file.