Org-mode now support multiple src blocks.
Previously, the lentic required matched blocks for "begin_src blah" with
"end_src". This fails if source blocks are used for more than one
language.

Here we add generic support for "unmatched" delimiters, where the close
block is identified as the next one to occur after an open. This also
means that the buffer cannot become invalid.

Closes #12.