Enable use of tags in org-mode src blocks.
Previously, org-mode src blocks could not contain tags which is
unfortunate as these are used to control many aspects of org-mode
functionality.

This patch supports tags by simply matching anything after the
language definition in the source block.

It also adds a strictly unrelated function to lentic-dev, which was
useful for debugging this feature.