Widening was inconsistent in hooks.
Lentic was widen buffers inconsistently in hooks, in particular the
before-command-function was not widened. This has been fixed a new macro
introduced which widens both buffers in a conf cleanly.

In this process of fixing this bug, I found that the orgel-org
configuration was dependent on the global current-buffer, rather than
obeying its own state; this was caused by a bug in m-buffer, hence the
dependency for this has been updated.

This also addresses a problem with "comment-region", in clojure buffers.
This uses narrowing to limit the effect of the commenting function. The
before-command-function was running with narrowing in place, which mean
that the location of the change was incorrectly calculated.

Closes #27