lentic-block changed to lentic-chunk.
lentic-block is really rather mis-named since "block" means the blocks
of the code that need to be commented and those that don't, as opposed
to "block comments" rather than "line comments". Renaming is a pain but
I want to support block comments so not doing this is going to make
things worse.