Add anchors to source comment blocks.
Source comment blocks are anchored to start of line in one direction,
but not in the other. Here the anchoring is made symmetrical, which
should prevent spurious matches to source blocks (occuring in strings,
for example, in org-mode source buffers).

This is an attempt to address issue #36, although without being able to
replicate the bug.