Regularisation of start/stop handling.
All configurations now use the same regexp to identify the start and
stop of comment sections, with optional comment block, and anchored to
the start of the line.

Previously, I had three different regexps with different variations of
anchors and groups. Nasty.

The anchoring should prevent spurious matches to source blocks (occuring
in strings, for example, in org-mode source buffers).

Closes issue #36.