Support for external publishing, add include dir
We now publish the lenticized documentation of lentic to an external
directory as this lowers the barrier to entry. Add support for single
CSS file for all lentic documentation, and include org-info.js.

Add an org-mode exporter which hides certain headers, and injects CSS
automatically.

Add support for additional files to be served by lentic server.